Add MemberList support for KCP
